// src/groups/item/team/operations/item/teamsAsyncOperationItemRequestBuilderPatchRequestConfiguration.ts
import {RequestOption} from '@microsoft/kiota-abstractions';

export interface TeamsAsyncOperationItemRequestBuilderPatchRequestConfiguration {
    /**
     * Request headers
     */
    headers?: Record<string, string[]> | undefined;
    /**
     * Request options
     */
    options?: RequestOption[] | undefined;
}